Update README.md
Dataset  is an Excel File
The dataset has two attributes, EV1 and EV2, and one output, Status.

If Status==1, the engine is running fine and needs no further analysis. 
If Status==0, the engine is operating in diminished capacity and engineers should perform a more detailed analysis to determine remaining engine life.